Moved crud operations to query runner
