+++ src/BlazorApp.Bootstrap.Ui/BlazorApp.Bootstrap.Ui.csproj
@@ -1,19 +1,20 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net9.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="MediatR" Version="12.5.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.4" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="MediatR" Version="12.5.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.4" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\BlazorApp.Bootstrap.Business\BlazorApp.Bootstrap.Business.csproj" />
-    <ProjectReference Include="..\BlazorApp.Bootstrap.Data\BlazorApp.Bootstrap.Data.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\BlazorApp.Bootstrap.Business\BlazorApp.Bootstrap.Business.csproj" />
+		<ProjectReference Include="..\BlazorApp.Bootstrap.Data\BlazorApp.Bootstrap.Data.csproj" />
+	</ItemGroup>
 
 </Project>
